requirements/local.txt
-r base.txt

factory-boy==2.7.0
fake-factory==0.7.2
django-debug-toolbar==1.5
sqlparse==0.2.1
